server/holidays/pyproject.toml
[tool.poetry]
name = "holidays"
version = "0.1.0"
description = ""
authors = ["Shubham Sudame <shubham.sudame@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.26.0"
python-dotenv = "^1.0.0"
beanie = "^1.24.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
